Relationships (#27)
* Relationships Update

Edited Driver.java Relationships.java and UML. java to handle the updates to the relationship object.

Relationships now have a source UML object, destination UML object and a type that is stored as a string.
The relationship class checks to make sure the type is one of the four allowed.

* ChangeRelType

Added a case in Driver.java to change a relationships type. Also changed rels in UML.java to a static.

* Relationships update

moved add relationship and delete relationship.

Also fixed other relationship methods.